fast_agend/pyproject.toml
[tool.poetry]
name = "fast-agend"
version = "0.1.0"
description = ""
authors = ["devgabriel <gabriel_fabri13@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.128.0"}
sqlalchemy = "^2.0.45"
psycopg = "^3.3.2"
alembic = "^1.17.2"
psycopg2-binary = "^2.9.11"
passlib = {version = "1.7.4", extras = ["bcrypt"]}
bcrypt = "3.2.2"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-dotenv = "^1.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
taskipy = "^1.14.1"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev fast_agend/app.py'
test = 'pytest -s -x --cov=fast_agend -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
